CDAP-13280 Add Spark 2 streaming Kafka source
