+++ kafka-plugins-0.10/docs/Kafka-alert-publisher.md
@@ -4,7 +4,7 @@
 
 Kafka Alert Publisher that allows you to publish alerts to kafka as json objects. The plugin internally uses kafka producer apis to publish alerts. 
 The plugin allows to specify kafka topic to use for publishing and other additional kafka producer properties. 
-This plugin uses kafka 0.8.2 java apis.
+This plugin uses kafka 0.10.2 java apis.
 
 Build
 -----
