+++ kafka-plugins-0.10/docs/KafkaAlerts-alertpublisher.md
@@ -6,7 +6,7 @@ Description
 Kafka Alert Publisher that allows you to publish alerts to kafka as json objects.
 The plugin internally uses kafka producer apis to publish alerts. 
 The plugin allows to specify kafka topic to use for publishing and other additional
-kafka producer properties. This plugin uses kafka 0.8.2 java apis.
+kafka producer properties. This plugin uses kafka 0.10.2 java apis.
 
 
 Configuration
@@ -16,6 +16,10 @@ Configuration
 **topic:** The Kafka topic to write to. This topic should already exist in kafka.
 
 **producerProperties** Specifies additional kafka producer properties like acks, client.id as key and value pair.
+
+**Kerberos Principal** The kerberos principal used for the source when kerberos security is enabled for kafka.
+
+**Keytab Location** The keytab location for the kerberos principal when kerberos security is enabled for kafka.
 
 Example
 -------
