+++ kafka-plugins-0.10/docs/kafka-batch-source-plugins-config.png
