+++ kafka-plugins-0.10/docs/kafka-sink-plugin-config.png
