+++ kafka-plugins-0.10/src/main/java/co/cask/hydrator/plugin/alertpublisher/KafkaAlertPublisher.java
@@ -29,11 +29,11 @@ import co.cask.cdap.etl.api.PipelineConfigurer;
 import co.cask.hydrator.common.KeyValueListParser;
 import com.google.common.base.Strings;
 import com.google.gson.Gson;
-import kafka.common.InvalidTopicException;
 import kafka.common.Topic;
 import org.apache.kafka.clients.producer.KafkaProducer;
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.clients.producer.ProducerRecord;
+import org.apache.kafka.common.errors.InvalidTopicException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -126,6 +126,16 @@ public class KafkaAlertPublisher extends AlertPublisher {
     @Nullable
     @Description("Additional kafka producer properties to set.")
     private String producerProperties;
+
+    @Description("The kerberos principal used for the source when kerberos security is enabled for kafka.")
+    @Macro
+    @Nullable
+    private String principal;
+
+    @Description("The keytab location for the kerberos principal when kerberos security is enabled for kafka.")
+    @Macro
+    @Nullable
+    private String keytabLocation;
 
     public Config(String brokers, String topic, String producerProperties) {
       this.brokers = brokers;
