+++ kafka-plugins-0.10/src/main/java/co/cask/hydrator/plugin/batchSource/KafkaSplit.java
