+++ kafka-plugins-0.10/widgets/Kafka-batchsource.json
