+++ kafka-plugins-0.10/widgets/Kafka-streamingsource.json
@@ -1,0 +1,151 @@
+{
+  "metadata": {
+    "spec-version": "1.5"
+  },
+  "display-name": "Kafka Consumer",
+  "configuration-groups": [
+    {
+      "label": "Kafka Configuration",
+      "properties": [
+        {
+          "widget-type": "textbox",
+          "label": "Reference Name",
+          "name": "referenceName"
+        },
+        {
+          "widget-type": "csv",
+          "label": "Kafka Brokers",
+          "name": "brokers",
+          "widget-attributes": {
+            "delimiter": ","
+          }
+        },
+        {
+          "widget-type": "textbox",
+          "label": "Kafka Topic",
+          "name": "topic"
+        },
+        {
+          "widget-type": "csv",
+          "label": "Topic Partitions",
+          "name": "partitions",
+          "widget-attributes": {
+            "delimiter": ","
+          }
+        },
+        {
+          "widget-type": "textbox",
+          "label": "Default Initial Offset",
+          "name": "defaultInitialOffset"
+        },
+        {
+          "widget-type": "keyvalue",
+          "label": "Initial Partition Offsets",
+          "name": "initialPartitionOffsets",
+          "widget-attributes": {
+            "showDelimiter": "false",
+            "key-placeholder": "Partition",
+            "value-placeholder": "Offset"
+          }
+        },
+        {
+          "widget-type": "textbox",
+          "label": "Time Field",
+          "name": "timeField"
+        },
+        {
+          "widget-type": "textbox",
+          "label": "Key Field",
+          "name": "keyField"
+        },
+        {
+          "widget-type": "textbox",
+          "label": "Partition Field",
+          "name": "partitionField"
+        },
+        {
+          "widget-type": "textbox",
+          "label": "Offset Field",
+          "name": "offsetField"
+        },
+        {
+          "widget-type": "textbox",
+          "label": "Max Rate Per Partition",
+          "name": "maxRatePerPartition",
+          "widget-attributes": {
+            "default": "1000"
+          }
+        },
+        {
+          "widget-type": "keyvalue",
+          "label": "Additional Kafka Consumer Properties",
+          "name": "kafkaProperties",
+          "widget-attributes": {
+            "showDelimiter": "false",
+            "key-placeholder": "Kafka consumer property",
+            "value-placeholder": "Kafka consumer property value"
+          }
+        }
+      ]
+    },
+    {
+      "label": "Format",
+      "properties": [
+        {
+          "widget-type": "select",
+          "label": "Format",
+          "name": "format",
+          "widget-attributes": {
+            "values": [
+              "",
+              "avro",
+              "binary",
+              "clf",
+              "csv",
+              "grok",
+              "syslog",
+              "text",
+              "tsv"
+            ],
+            "default": ""
+          }
+        }
+      ]
+    },
+    {
+      "label": "Authentication",
+      "properties": [
+        {
+          "widget-type": "textbox",
+          "label": "Kerberos Principal",
+          "name": "principal"
+        },
+        {
+          "widget-type": "textbox",
+          "label": "Keytab Location",
+          "name": "keytabLocation"
+        }
+      ]
+    }
+  ],
+  "outputs": [
+    {
+      "name": "schema",
+      "widget-type": "schema",
+      "widget-attributes": {
+        "default-schema": {
+          "name": "etlSchemaBody",
+          "type": "record",
+          "fields": [
+            {
+              "name": "message",
+              "type": ["bytes", "null"]
+            }
+          ]
+        },
+        "schema-default-type": "string",
+        "property-watch": "format"
+      }
+    }
+  ]
+}
