+++ kafka-plugins-0.10/widgets/KafkaAlerts-alertpublisher.json
@@ -1,0 +1,52 @@
+{
+  "metadata": {
+    "spec-version": "1.5"
+  },
+  "display-name": "Kafka Alert Publisher",
+  "configuration-groups": [
+    {
+      "label": "Kafka Alert Publisher Config",
+      "properties": [
+        {
+          "widget-type": "csv",
+          "label": "Kafka Brokers",
+          "name": "brokers",
+          "widget-attributes": {
+            "delimiter": ","
+          }
+        },
+        {
+          "widget-type": "textbox",
+          "label": "Kafka Topic",
+          "name": "topic"
+        },
+        {
+          "widget-type": "keyvalue",
+          "label": "Additional Kafka Producer Properties",
+          "name": "producerProperties",
+          "widget-attributes": {
+            "showDelimiter": "false",
+            "key-placeholder": "Kafka producer property",
+            "value-placeholder": "Kafka producer property value"
+          }
+        }
+      ]
+    },
+    {
+      "label": "Authentication",
+      "properties": [
+        {
+          "widget-type": "textbox",
+          "label": "Kerberos Principal",
+          "name": "principal"
+        },
+        {
+          "widget-type": "textbox",
+          "label": "Keytab Location",
+          "name": "keytabLocation"
+        }
+      ]
+    }
+  ],
+  "outputs": [ ]
+}
