+++ kafka-plugins-0.8/src/test/java/co/cask/hydrator/KafkaStreamingSourceTest.java
@@ -1,0 +1,268 @@
+/*
+ * Copyright © 2018 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package co.cask.hydrator;
+
+import co.cask.cdap.api.artifact.ArtifactRange;
+import co.cask.cdap.api.artifact.ArtifactSummary;
+import co.cask.cdap.api.artifact.ArtifactVersion;
+import co.cask.cdap.api.data.format.StructuredRecord;
+import co.cask.cdap.api.data.schema.Schema;
+import co.cask.cdap.api.dataset.table.Table;
+import co.cask.cdap.common.utils.Networks;
+import co.cask.cdap.common.utils.Tasks;
+import co.cask.cdap.datapipeline.DataPipelineApp;
+import co.cask.cdap.datastreams.DataStreamsApp;
+import co.cask.cdap.datastreams.DataStreamsSparkLauncher;
+import co.cask.cdap.etl.api.streaming.StreamingSource;
+import co.cask.cdap.etl.mock.batch.MockSink;
+import co.cask.cdap.etl.mock.test.HydratorTestBase;
+import co.cask.cdap.etl.proto.v2.DataStreamsConfig;
+import co.cask.cdap.etl.proto.v2.ETLPlugin;
+import co.cask.cdap.etl.proto.v2.ETLStage;
+import co.cask.cdap.proto.artifact.AppRequest;
+import co.cask.cdap.proto.id.ApplicationId;
+import co.cask.cdap.proto.id.ArtifactId;
+import co.cask.cdap.proto.id.NamespaceId;
+import co.cask.cdap.test.ApplicationManager;
+import co.cask.cdap.test.DataSetManager;
+import co.cask.cdap.test.SparkManager;
+import co.cask.cdap.test.TestConfiguration;
+import co.cask.hydrator.common.http.HTTPPollConfig;
+import co.cask.hydrator.plugin.source.KafkaStreamingSource;
+import com.google.common.base.Charsets;
+import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.ImmutableSet;
+import com.google.common.util.concurrent.Uninterruptibles;
+import kafka.common.TopicAndPartition;
+import kafka.serializer.DefaultDecoder;
+import org.apache.spark.streaming.kafka.KafkaUtils;
+import org.apache.twill.internal.kafka.EmbeddedKafkaServer;
+import org.apache.twill.internal.kafka.client.ZKKafkaClientService;
+import org.apache.twill.internal.zookeeper.InMemoryZKServer;
+import org.apache.twill.kafka.client.Compression;
+import org.apache.twill.kafka.client.KafkaClientService;
+import org.apache.twill.kafka.client.KafkaPublisher;
+import org.apache.twill.zookeeper.ZKClientService;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.ClassRule;
+import org.junit.Test;
+import org.junit.rules.TemporaryFolder;
+
+import java.io.File;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Properties;
+import java.util.Set;
+import java.util.concurrent.Callable;
+import java.util.concurrent.TimeUnit;
+
+/**
+ * Tests for Spark plugins.
+ */
+public class KafkaStreamingSourceTest extends HydratorTestBase {
+
+  @ClassRule
+  public static final TestConfiguration CONFIG = new TestConfiguration("explore.enabled", false);
+
+  private static final ArtifactId DATAPIPELINE_ARTIFACT_ID =
+    NamespaceId.DEFAULT.artifact("data-pipeline", "4.3.2");
+  private static final ArtifactId DATASTREAMS_ARTIFACT_ID =
+    NamespaceId.DEFAULT.artifact("data-streams", "4.3.2");
+  private static final ArtifactSummary DATASTREAMS_ARTIFACT =
+    new ArtifactSummary("data-streams", "4.3.2");
+
+  private static ZKClientService zkClient;
+  private static KafkaClientService kafkaClient;
+  private static InMemoryZKServer zkServer;
+  private static EmbeddedKafkaServer kafkaServer;
+  private static int kafkaPort;
+
+  @ClassRule
+  public static TemporaryFolder tmpFolder = new TemporaryFolder();
+
+
+  @BeforeClass
+  public static void setupTest() throws Exception {
+    // add the artifact for data pipeline app
+    setupBatchArtifacts(DATAPIPELINE_ARTIFACT_ID, DataPipelineApp.class);
+
+    setupStreamingArtifacts(DATASTREAMS_ARTIFACT_ID, DataStreamsApp.class);
+
+    // add artifact for spark plugins
+    Set<ArtifactRange> parents = ImmutableSet.of(
+      new ArtifactRange(NamespaceId.DEFAULT.getNamespace(), DATAPIPELINE_ARTIFACT_ID.getArtifact(),
+                        new ArtifactVersion(DATAPIPELINE_ARTIFACT_ID.getVersion()), true,
+                        new ArtifactVersion(DATAPIPELINE_ARTIFACT_ID.getVersion()), true),
+      new ArtifactRange(NamespaceId.DEFAULT.getNamespace(), DATASTREAMS_ARTIFACT_ID.getArtifact(),
+                        new ArtifactVersion(DATASTREAMS_ARTIFACT_ID.getVersion()), true,
+                        new ArtifactVersion(DATASTREAMS_ARTIFACT_ID.getVersion()), true)
+    );
+    addPluginArtifact(NamespaceId.DEFAULT.artifact("spark-plugins", "1.0.0"), parents,
+                      KafkaStreamingSource.class, KafkaUtils.class, DefaultDecoder.class, TopicAndPartition.class,
+                      HTTPPollConfig.class);
+
+    zkServer = InMemoryZKServer.builder().setDataDir(TMP_FOLDER.newFolder()).build();
+    zkServer.startAndWait();
+
+    kafkaPort = Networks.getRandomPort();
+    kafkaServer = new EmbeddedKafkaServer(generateKafkaConfig(zkServer.getConnectionStr(),
+                                                              kafkaPort, TMP_FOLDER.newFolder()));
+    kafkaServer.startAndWait();
+
+    zkClient = ZKClientService.Builder.of(zkServer.getConnectionStr()).build();
+    zkClient.startAndWait();
+
+    kafkaClient = new ZKKafkaClientService(zkClient);
+    kafkaClient.startAndWait();
+  }
+
+  @AfterClass
+  public static void cleanup() {
+    kafkaClient.stopAndWait();
+    kafkaServer.stopAndWait();
+    zkClient.stopAndWait();
+    zkServer.stopAndWait();
+  }
+
+  @Test
+  public void testKafkaStreamingSource() throws Exception {
+    Schema schema = Schema.recordOf(
+      "user",
+      Schema.Field.of("id", Schema.of(Schema.Type.LONG)),
+      Schema.Field.of("first", Schema.of(Schema.Type.STRING)),
+      Schema.Field.of("last", Schema.of(Schema.Type.STRING)));
+
+    Map<String, String> properties = new HashMap<>();
+    properties.put("referenceName", "kafkaPurchases");
+    properties.put("brokers", "localhost:" + kafkaPort);
+    properties.put("topic", "users");
+    properties.put("defaultInitialOffset", "-2");
+    properties.put("format", "csv");
+    properties.put("schema", schema.toString());
+
+    ETLStage source = new ETLStage("source", new ETLPlugin("Kafka", StreamingSource.PLUGIN_TYPE, properties, null));
+
+    DataStreamsConfig etlConfig = DataStreamsConfig.builder()
+      .addStage(source)
+      .addStage(new ETLStage("sink", MockSink.getPlugin("kafkaOutput")))
+      .addConnection("source", "sink")
+      .setBatchInterval("1s")
+      .setStopGracefully(true)
+      .build();
+
+    AppRequest<DataStreamsConfig> appRequest = new AppRequest<>(DATASTREAMS_ARTIFACT, etlConfig);
+    ApplicationId appId = NamespaceId.DEFAULT.app("KafkaSourceApp");
+    ApplicationManager appManager = deployApplication(appId, appRequest);
+
+    // write some messages to kafka
+    Map<String, String> messages = new HashMap<>();
+    messages.put("a", "1,samuel,jackson");
+    messages.put("b", "2,dwayne,johnson");
+    messages.put("c", "3,christopher,walken");
+    sendKafkaMessage("users", messages);
+
+    SparkManager sparkManager = appManager.getSparkManager(DataStreamsSparkLauncher.NAME);
+    sparkManager.start();
+    sparkManager.waitForStatus(true, 10, 1);
+
+    final DataSetManager<Table> outputManager = getDataset("kafkaOutput");
+    Tasks.waitFor(
+      ImmutableMap.of(1L, "samuel jackson", 2L, "dwayne johnson", 3L, "christopher walken"),
+      new Callable<Map<Long, String>>() {
+        @Override
+        public Map<Long, String> call() throws Exception {
+          outputManager.flush();
+          Map<Long, String> actual = new HashMap<>();
+          for (StructuredRecord outputRecord : MockSink.readOutput(outputManager)) {
+            actual.put((Long) outputRecord.get("id"), outputRecord.get("first") + " " + outputRecord.get("last"));
+          }
+          return actual;
+        }
+      },
+      2,
+      TimeUnit.MINUTES);
+
+    sparkManager.stop();
+    sparkManager.waitForStatus(false, 10, 1);
+
+    // clear the output table
+    MockSink.clear(outputManager);
+
+    // now write some more messages to kafka and start the program again to make sure it picks up where it left off
+    messages = new HashMap<>();
+    messages.put("d", "4,terry,crews");
+    messages.put("e", "5,sylvester,stallone");
+    sendKafkaMessage("users", messages);
+
+    sparkManager.start();
+    sparkManager.waitForStatus(true, 10, 1);
+
+    Tasks.waitFor(
+      ImmutableMap.of(4L, "terry crews", 5L, "sylvester stallone"),
+      new Callable<Map<Long, String>>() {
+        @Override
+        public Map<Long, String> call() throws Exception {
+          outputManager.flush();
+          Map<Long, String> actual = new HashMap<>();
+          for (StructuredRecord outputRecord : MockSink.readOutput(outputManager)) {
+            actual.put((Long) outputRecord.get("id"), outputRecord.get("first") + " " + outputRecord.get("last"));
+          }
+          return actual;
+        }
+      },
+      2,
+      TimeUnit.MINUTES);
+
+    sparkManager.stop();
+  }
+
+  private static Properties generateKafkaConfig(String zkConnectStr, int port, File logDir) {
+    Properties prop = new Properties();
+    prop.setProperty("log.dir", logDir.getAbsolutePath());
+    prop.setProperty("port", Integer.toString(port));
+    prop.setProperty("broker.id", "1");
+    prop.setProperty("num.partitions", "1");
+    prop.setProperty("zookeeper.connect", zkConnectStr);
+    prop.setProperty("zookeeper.connection.timeout.ms", "1000000");
+    prop.setProperty("default.replication.factor", "1");
+    return prop;
+  }
+
+  private void sendKafkaMessage(@SuppressWarnings("SameParameterValue") String topic, Map<String, String> messages) {
+    KafkaPublisher publisher = kafkaClient.getPublisher(KafkaPublisher.Ack.ALL_RECEIVED, Compression.NONE);
+
+    // If publish failed, retry up to 20 times, with 100ms delay between each retry
+    // This is because leader election in Kafka 08 takes time when a topic is being created upon publish request.
+    int count = 0;
+    do {
+      KafkaPublisher.Preparer preparer = publisher.prepare(topic);
+      for (Map.Entry<String, String> entry : messages.entrySet()) {
+        preparer.add(Charsets.UTF_8.encode(entry.getValue()), entry.getKey());
+      }
+      try {
+        preparer.send().get();
+        break;
+      } catch (Exception e) {
+        // Backoff if send failed.
+        Uninterruptibles.sleepUninterruptibly(100, TimeUnit.MILLISECONDS);
+      }
+    } while (count++ < 20);
+  }
+
+
+}
