--- kafka-plugins-0.9/pom.xml
@@ -1,78 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <parent>
-    <artifactId>kafka-plugins</artifactId>
-    <groupId>co.cask.hydrator</groupId>
-    <version>1.8.1</version>
-  </parent>
-  <modelVersion>4.0.0</modelVersion>
-
-  <name>Apache Kafka 0.9 plugins</name>
-  <artifactId>kafka-plugins</artifactId>
-  <version>1.8.1-0.9.0.1</version>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.apache.kafka</groupId>
-      <artifactId>kafka_2.10</artifactId>
-      <version>${kafka9.version}</version>
-      <exclusions>
-        <exclusion>
-          <groupId>org.slf4j</groupId>
-          <artifactId>slf4j-log4j12</artifactId>
-        </exclusion>
-      </exclusions>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.felix</groupId>
-        <artifactId>maven-bundle-plugin</artifactId>
-        <version>3.3.0</version>
-        <extensions>true</extensions>
-        <configuration>
-          <instructions>
-            <_exportcontents>co.cask.hydrator.plugin.*;org.apache.spark.streaming.kafka.*;
-              kafka.serializer.*;kafka.common;org.apache.kafka.common.serialization.*;
-              org.apache.kafka.clients.consumer.*</_exportcontents>
-            <Embed-Dependency>*;inline=false;scope=compile</Embed-Dependency>
-            <Embed-Transitive>true</Embed-Transitive>
-            <Embed-Directory>lib</Embed-Directory>
-          </instructions>
-        </configuration>
-        <executions>
-          <execution>
-            <phase>package</phase>
-            <goals>
-              <goal>bundle</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-      <plugin>
-        <groupId>co.cask</groupId>
-        <artifactId>cdap-maven-plugin</artifactId>
-        <version>1.0.0</version>
-        <configuration>
-          <cdapArtifacts>
-            <parent>system:cdap-data-pipeline[4.3.0-SNAPSHOT,6.0.0-SNAPSHOT)</parent>
-          </cdapArtifacts>
-        </configuration>
-        <executions>
-          <execution>
-            <id>create-artifact-config</id>
-            <phase>prepare-package</phase>
-            <goals>
-              <goal>create-plugin-json</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins>
-  </build>
-
-</project>
